docs: computer player
